Add score submission as runtime param

+++ primitives/src/lib.rs
@@ -121,7 +121,7 @@ pub const DEFAULT_SESSIONS_PER_ERA: SessionIndex = 3;
 
 /// Gather ABFT score every SCORE_SUBMISSION_PERIOD rounds
 #[cfg(feature = "short_session")]
-pub const SCORE_SUBMISSION_PERIOD: usize = 15;
+pub const SCORE_SUBMISSION_PERIOD: u32 = 15;
 // --------------- Test build end  ---------------------
 
 // --------------- Production build ---------------------
@@ -135,7 +135,7 @@ pub const DEFAULT_SESSIONS_PER_ERA: SessionIndex = 96;
 
 /// Gather ABFT score every SCORE_SUBMISSION_PERIOD rounds
 #[cfg(not(feature = "short_session"))]
-pub const SCORE_SUBMISSION_PERIOD: usize = 300;
+pub const SCORE_SUBMISSION_PERIOD: u32 = 300;
 // --------------- Production build end ---------------------
 
 /// How many decimals AZERO coin has
